+++ src/settings/components/time-picker.jsx
@@ -7,11 +7,11 @@ import React from 'react';
 
 import { Container } from '@zextras/carbonio-design-system';
 import { t } from '@zextras/carbonio-shell-ui';
-import momentLocalizer from 'react-widgets-moment';
 
 import StartTimePicker from './start-time-picker';
+import { dateFnsLocalizer } from '../../commons/date-fns-react-widgets-localizer';
 
-momentLocalizer();
+dateFnsLocalizer();
 
 export default function DatePicker({ start, end, onChange, day, disabled }) {
 	return (