+++ src/shared/invite-response/invite-response.tsx
@@ -9,10 +9,9 @@ import styled from '@emotion/styled';
 import { Container, Row, Icon, Divider, Spinner } from '@zextras/carbonio-design-system';
 import { useUserAccount } from '@zextras/carbonio-shell-ui';
 import { ROOT_NAME, FOLDERS, getRootAccountId, useRoot } from '@zextras/carbonio-ui-commons';
+import { endOfDay } from 'date-fns';
 import { filter, includes } from 'lodash';
-import moment from 'moment';
 
-import 'moment-timezone';
 import { AvailabilityChecker } from './parts/availability-checker';
 import { EventDetails } from './parts/event-details';
 import InviteHeaderPart from './parts/invite-header-part';
@@ -25,6 +24,7 @@ import { MESSAGE_METHOD } from '../../constants/api';
 import { normalizeInvite } from '../../normalizations/normalize-invite';
 import { StoreProvider } from '../../store/redux';
 import type { InviteResponseArguments } from '../../types/integrations';
+import { parseDateFromICS } from '../../utils/dates';
 
 const InviteContainer = styled(Container)`
 	border: 0.0625rem solid ${({ theme }): string => theme.palette.gray2.regular};
@@ -66,7 +66,8 @@ export const InviteResponse: FC<InviteResponseArguments> = ({
 		[mailMsg?.participants]
 	);
 
-	const getEndOfDay = (day: string | number): number => moment(day).endOf('day').valueOf();
+	const getEndOfDay = (day: string | number): number =>
+		endOfDay(typeof day === 'string' ? parseDateFromICS(day) : new Date(day)).getTime();
 
 	const proposedStartTime = mailMsg.invite[0]?.comp?.[0]?.s?.[0]?.d;
 	const proposedEndTime = mailMsg.invite[0]?.comp?.[0]?.e?.[0]?.d;
@@ -98,8 +99,11 @@ export const InviteResponse: FC<InviteResponseArguments> = ({
 					<AvailabilityChecker
 						email={email}
 						rootId={root.id}
-						start={invite?.start?.u ?? moment(proposedStartTime).valueOf()}
-						end={invite?.end?.u ?? getEndOfDay(proposedEndTime)}
+						start={
+							invite?.start?.u ??
+							(proposedStartTime ? parseDateFromICS(proposedStartTime).getTime() : 0)
+						}
+						end={invite?.end?.u ?? (proposedEndTime ? getEndOfDay(proposedEndTime) : 0)}
 						allDay={invite.allDay ?? false}
 						uid={invite.uid}
 					/>
@@ -108,9 +112,13 @@ export const InviteResponse: FC<InviteResponseArguments> = ({
 					<ProposedTimeReply
 						id={invite?.apptId}
 						start={
-							proposedStartTime ? moment(proposedStartTime).valueOf() : (invite?.start?.u ?? 0)
+							proposedStartTime
+								? parseDateFromICS(proposedStartTime).getTime()
+								: (invite?.start?.u ?? 0)
 						}
-						end={proposedEndTime ? moment(proposedEndTime).valueOf() : (invite?.end?.u ?? 0)}
+						end={
+							proposedEndTime ? parseDateFromICS(proposedEndTime).getTime() : (invite?.end?.u ?? 0)
+						}
 						moveToTrash={moveToTrash}
 						title={mailMsg.subject}
 						to={to}