+++ src/view/editor/editor-util-hooks/use-get-public-url.jsx
@@ -8,8 +8,8 @@ import { useCallback, useMemo } from 'react';
 
 import { useSnackbar } from '@zextras/carbonio-design-system';
 import { t, useIntegratedFunction } from '@zextras/carbonio-shell-ui';
+import { format } from 'date-fns';
 import { filter, map } from 'lodash';
-import moment from 'moment';
 
 import { useAppDispatch, useAppSelector } from '../../../store/redux/hooks';
 import {
@@ -31,7 +31,7 @@ export const useGetPublicUrl = ({ editorId }) => {
 		() =>
 			t('label.public_link_description', {
 				title,
-				date: moment().format('yyyy-MM-DD_hh:mm'),
+				date: format(new Date(), 'yyyy-MM-dd_HH:mm'),
 				defaultValue: 'Generated from {{title}} on {{date}}'
 			}),
 		[title]